converted ReviewListModal, adjustment to prop typing in ReviewDisplay
